Nagios speed optimizations
This patch set introduces per-host and per-hostgroup config files,
replacing the monolithic charm.cfg previously used to manage the
hosts/hostgroups/services of the NRPE peers.  It also tries to
minimize the number of such files which are rewritten on a particular
hook execution.

Migration from charm.cfg to the new format is automatic.

This changeset also introduces a rewrite-peer-config action.  This is
in case of unexpected misses regarding rewriting of peer files,
unexpected corner cases that the previous "rewrite everything"
methodology may have covered up, and as a method of accomodating for
hooks which may have not fired correctly.  This action will delete
charm.cfg and all host_*.cfg and hostgroup_*.cfg files, effectively
mimicking the "rewrite everything" logic, albeit to the individual
files.