Adding cleanup logic
This handles a corner case where the last remaining host of a
duplicate set ends up with multiple host files due to the way that the
code minimizes writing of host files.  Any hosts which lack files in
expected locations will have them created, but removal of existing
files is only done for units directly related to a relation hook - thus
hosts which are renamed due to addition/removal of a different unit
are having new files written but not necessarily having their old files
removed.