Keep writing config for newly-created config files
With the new logic, we try to write only the files we need to.
However, there may legitimately be cases where there's multiple sets
of data associated with the same host name (e.g. multiple nrpe units),
and at present, only the first set of such data will actually be
written.

This change tracks files which are being newly created, and will allow
multiple sets of relation data to be written to the same file, if
necessary.