Allow for tuning of notification interval
Notification interval is handled at the level of the base
generic-service and generic-host objects.  generic-service and
generic-host are not defined by the charm, but rather by the installed
nagios3 package.  This patch uses PyNag to tweak those values, similar
to how the update_localhost function works for tweaking the localhost
host object.