Allow for limiting email notifications
By setting email_max_notifications to a positive value, email
notifications will be sent only up to N times per alert, after which
no more mails will be sent.

This is intended to be used in conjunction with a non-zero
notification_interval, allowing for alerts to re-notify but without
sending emails on each re-notification.