Add config option for nrpe packet version
When using ssl in nagios, we get error as shown below

Request packet version was invalid!

This is due to the fact it is usig v3 first, and then v2. This patch
ensures that we can change the behavior if needed. In this scenario
where this is an issue, we get the error below. The config to use
version 2 prevents this

nrpe[2383494]: Error: (use_ssl == true): Request packet version was invalid!
nrpe[2383494]: Could not read request from client 10.0.1.121, bailing out...
nrpe[2383494]: INFO: SSL Socket Shutdown.